refactor: code review の指摘に対応 (壊れた rules の隔離・書き込み時検証・型側のみの readonly)
- listPlaylists は壊れた rules の行を rules なしで返して警告を残し、存在確認は SELECT 1 の assertPlaylistExists に分離して削除・改名で復旧できるようにする
- createPlaylist / updatePlaylist は smartPlaylistRulesSchema で parse してから保存し、updatePlaylist は COALESCE で現在行を読まずに更新する
- decade の .int() を外し year と同じ number にする
- readonly は types.ts の DeepReadonly で型側だけに付け、zod の .readonly() (実行時の Object.freeze) をやめる
- audioFormatSchema から mme への依存を外し、MusicRowInput.audioFormat の型で mme の全フォーマットが列挙に含まれることを検証する
- types.ts のスキーマ参照パス、database.md の版表記 (v1.2)、createPlaylist の行スキーマ定数化、orphanRowSchema の idRowSchema 継承

+++ src/main/library/getFilterOptions.ts
@@ -4,9 +4,13 @@ import { genreCountSchema } from "../../shared/schemas/filterOptionsSchema";
 import type { FilterOptions } from "../ipc/types";
 import { ALBUM_ARTIST_SQL } from "./ALBUM_ARTIST_SQL";
 
-/** Row of the decade query; a NULL bucket holds the unknown-year albums. */
+/**
+ * Row of the decade query; a NULL bucket holds the unknown-year albums. Like
+ * `musicSchema.year`, the decade is a plain number: integer-ness is a business
+ * rule of the write path, not of the column.
+ */
 const decadeBucketSchema = z.object({
-  decade: z.number().int().nullable(),
+  decade: z.number().nullable(),
   count: z.number().int(),
 });
